requirements.txt
aiomysql==0.0.20
tortoise-orm==0.12.0
sanic==19.3.1
sanic-cookiesession==0.2.0
sanic-jinja2==0.7.5
werkzeug==0.15.4
pyjwt==1.7.1
pyquery==1.4.0
aioredis==1.2.0
pyOpenSSL==19.0.0
pyasn1==0.4.7